% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsSeverin.R
\name{blotIt_out_to_list}
\alias{blotIt_out_to_list}
\title{blotIt out to list converter}
\usage{
blotIt_out_to_list(out, use_factors = T)
}
\arguments{
\item{out}{'out' file as produced by \code{alignME()}.}

\item{use_factors}{Logical, if set to \code{TRUE} the name columns are converted to
factors. By this, the plotted order of target will be same order as they are
in the source data file fad through \link{read.wide} to \link{alignME}.}
}
\description{
Takes the output of blotIt2 and converts it to a list by reading out the
attributes.
}
