% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsSvenja.R
\name{getReplicates}
\alias{getReplicates}
\title{Get scaled replicates}
\usage{
getReplicates(out)
}
\arguments{
\item{out}{output of \link{alignME}}
}
\description{
Extract scaled replicates from the blotIt output.
}
\author{
Svenja Kemmer
}
