% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MAPK_raw}
\alias{MAPK_raw}
\title{Time-course data for the MAPK cell signaling pathway}
\description{
Phosphorylated ERK1 and MEK have been measured at times 0, ..., 60.
"Experiment" refers to the western blot gel from which the
measurement originates whereas "Condition" refers to the
stimulation or treatment. Wide format as frequently found in a lab.
}
\keyword{data}

