% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot1}
\alias{plot1}
\title{Plot 1}
\usage{
plot1(out, ..., residual = FALSE)
}
\arguments{
\item{out}{output of \link{alignME}.}

\item{...}{logical expression used for subsetting the data frames,
e.g. \code{name == "pERK1" & time < 60}.}
}
\description{
Plot original data together with the prediction of the alignment
model.
}
\seealso{
\link{plot2}, \link{plot3}, \link{plot4}
}

