% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot2}
\alias{plot2}
\title{Plot 2}
\usage{
plot2(out, ...)
}
\arguments{
\item{out}{output of \link{alignME}.}

\item{...}{logical expression used for subsetting the data frames,
e.g. \code{name == "pERK1" & time < 60}.}
}
\description{
Plot the prediction of the alignment model together with the
original data transformed to the scale of the predicted
time-course.
}
\seealso{
\link{plot1}, \link{plot3}, \link{plot4}
}

