% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot3}
\alias{plot3}
\title{Plot 3}
\usage{
plot3(out, ...)
}
\arguments{
\item{out}{output of \link{alignME}.}

\item{...}{logical expression used for subsetting the data frames,
e.g. \code{name == "pERK1" & time < 60}.}
}
\description{
Plot the prediction of the alignment model grouped by fixed
effects.
}
\seealso{
\link{plot1}, \link{plot2}, \link{plot4}
}

