% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsSeverin.R
\name{plot_blotIt}
\alias{plot_blotIt}
\title{All-in-one plot function for blotIt()}
\usage{
plot_blotIt(
  out,
  plot_points = "original",
  plot_line = "prediction",
  spline = F,
  scales = "fixed",
  plot_caption = T,
  ncol = NULL,
  ...
)
}
\arguments{
\item{out}{'out' file as produced by 'alignME()'. Either as raw output or converted
to a list, e.g. per \link{blotIt_out_to_list}.}

\item{plot_points}{String to specify which data set should be plotted in form of points
with corresponding error bars. It must be one of \code{c("original", "scaled", "prediction", "aligned")}.}

\item{plot_line}{Same as above but with a line and errorband.}

\item{spline}{Logical, if set to \code{TRUE}, what is specified as \code{plot_line} will be plotted
as a smooth spline instead of straight lines between points.}

\item{scales}{String passed as \code{scales} argument to \link{facet_wrap}.}

\item{plot_caption}{Logical, if \code{TRUE}, a caption describing the plotted data is added to the plot.}

\item{ncol}{Numerical passed as \code{ncol} argument to \link{facet_wrap}.}

\item{...}{Logical expression used for subsetting the data frames, e.g. name == "pERK1" & time < 60.


\describe{
To reproduce the known function \link{plot1}, \link{plot2} and \link{plot3},
use:
\item{plot1}{\code{plot_points} = 'original', \code{plot_line} = 'prediction'}
\item{plot2}{\code{plot_points} = 'scaled', \code{plot_line} = 'aligned'}
\item{plot3}{\code{plot_points} = 'aligned', \code{plot_line} = 'aligned'}
}}
}
\description{
Takes the output of 'alignME()' If not already in a list it is internally converted
by us of \link{blotIt_out_to_list}. Which data will be plotted can then be
specefied seperately.
}
