% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsSvenja.R
\name{plotDR1}
\alias{plotDR1}
\title{PlotDR 1}
\usage{
plotDR1(out, ...)
}
\arguments{
\item{out}{output of \link{alignME} with columns time, dose, ...}

\item{...}{logical expression used for subsetting the data frames,
e.g. \code{name == "pERK1" & time < 60}.}
}
\description{
Plot original dose response data together with the prediction of the alignment
model in new style.
}
\seealso{
\link{plotNew1}, \link{plotNew2}, \link{plotNew3}, \link{plotDR2}, \link{plotDR3}
}
