% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsSvenja.R
\name{plotDR2}
\alias{plotDR2}
\title{PlotDR 2}
\usage{
plotDR2(out, ...)
}
\arguments{
\item{out}{output of \link{alignME} with columns time, dose, ...}

\item{...}{logical expression used for subsetting the data frames,
e.g. \code{name == "pERK1" & time < 60}.}
}
\description{
Plot the dose response prediction of the alignment model together with the
original data transformed to the scale of the predicted
time-course in new style.
}
\seealso{
\link{plotNew1}, \link{plotNew2}, \link{plotNew3}, \link{plotDR1}, \link{plotDR3}
}
