% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsSvenja.R
\name{plotDR3}
\alias{plotDR3}
\title{PlotDR 3}
\usage{
plotDR3(out, ...)
}
\arguments{
\item{out}{output of \link{alignME} with columns time, dose, ...}

\item{...}{logical expression used for subsetting the data frames,
e.g. \code{name == "pERK1" & time < 60}.}
}
\description{
Plot the dose response prediction of the alignment model grouped by fixed
effects in new style.
}
\seealso{
\link{plotNew1}, \link{plotNew2}, \link{plotNew3}, \link{plotDR1}, \link{plotDR2}
}
