% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsSvenja.R
\name{plotNew1}
\alias{plotNew1}
\title{PlotNew 1}
\usage{
plotNew1(out, ..., residual = FALSE)
}
\arguments{
\item{out}{output of \link{alignME}.}

\item{...}{logical expression used for subsetting the data frames,
e.g. \code{name == "pERK1" & time < 60}.}
}
\description{
Plot original data together with the prediction of the alignment
model in new style.
}
\seealso{
\link{plotNew2}, \link{plotNew3}, \link{plotDR1}, \link{plotDR2}, \link{plotDR3}
}
