% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blotIt2.R
\name{read.wide}
\alias{read.wide}
\title{Read data from csv or txt file}
\usage{
read.wide(file, description = 1:3, time = 1, header = TRUE, ...)
}
\arguments{
\item{file}{character, the name of the data file.}

\item{description}{numeric index vector of the colums containing
the description.}

\item{time}{numeric index of length 1, the time column.}

\item{header}{a logical value indicating whether the file contains
the names of the variables as its first line.}

\item{...}{further arguments being passed to \link{read.csv}.}
}
\value{
data frame with columns "name", "time", "value" and other
columns describing the measurements.
}
\description{
Reads data stored in a wide format, i.e. description
columns with time points, experimental conditions, etc. are
followed by each one column per measured target.
}
