% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blotIt2.R
\name{splitData}
\alias{splitData}
\title{splitData}
\usage{
splitData(data, fixed, latent, normalize_input, log)
}
\arguments{
\item{data}{data frame with columns "name", "time", "value" and others}

\item{fixed}{two-sided formula, see \link{alignME}}

\item{latent}{two-sided formula, see \link{alignME}}

\item{normalize_input}{logical if set to TRUE, the input data will normalized
per latent effect by dividing by the respective mean. Prevents convergence
failure on some hardware when the data for different latent effects differ by
to many orders of magnitude.}
}
\value{
list of data frames
}
\description{
Split data in independent blocks according to fixed and latent
variables as being defined for \link{alignME}.
}
