% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsSvenja.R
\name{plotDR2}
\alias{plotDR2}
\title{PlotDR 2}
\usage{
plotDR2(out, ..., ymin = NULL, units = c("", ""))
}
\arguments{
\item{out}{output of \link{alignME} with obligatory columns time and dose}

\item{...}{logical expression used for subsetting the data frames,
e.g. \code{name == "pERK1" & time < 60}.}

\item{ymin}{value defining ymin for all y-axes}

\item{units}{character vector of length 2 defining units for the x- and y-axes,
e.g. \code{c("[ng/ml]","[a.u.]")}}
}
\description{
Plot the dose response prediction of the alignment model together with the
original data transformed to the scale of the predicted
time-course in new style.
}
\seealso{
\link{plotNew1}, \link{plotNew2}, \link{plotNew3}, \link{plotDR1}, \link{plotDR3}
}
\author{
Svenja Kemmer
}
