% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blotIt2.R
\name{splitData}
\alias{splitData}
\title{splitData}
\usage{
splitData(data, fixed, latent)
}
\arguments{
\item{data}{data frame with columns "name", "time", "value" and others}

\item{fixed}{two-sided formula, see \link{alignME}}

\item{latent}{two-sided formula, see \link{alignME}}
}
\value{
list of data frames
}
\description{
Split data in independent blocks according to fixed and latent
variables as being defined for \link{alignME}.
}
